Corrections done

+++ src/main/java/ch/heigvd/dai/Main.java
@@ -1,77 +1,94 @@
 package ch.heigvd.dai;
+
 import java.util.Objects;
 import java.util.concurrent.Callable;
-
-import org.w3c.dom.Text;
 import picocli.CommandLine;
 
 @CommandLine.Command(
         name = "Find-Replace-Tool",
-        description = "Executes different action on an input text.",
+        description = "A small program to perform different action on an input text.",
         version = "1.0.0",
-        mixinStandardHelpOptions = true)
+        mixinStandardHelpOptions = true
+)
 
 class Main implements Callable<Integer> {
-
     @CommandLine.Option(
             names = {"-f", "--file"},
-            description = "The path of the file to read",
-            required = true
+            description = "The path of the file to read. Not usable if the parameter --text (-t) is used."
     ) private String input_filePath = "";
 
     @CommandLine.Option(
             names = {"-t", "--text"},
-            description = "The input text to read"
+            description = "The input text to read. Not usable if the parameter --file (-f) is used."
     ) private String input_terminalText = "";
 
     @CommandLine.Option(
             names = {"-o", "--output"},
-            description = "The path of the file to write in"
+            description = "The path of the output file to write in.",
+            defaultValue = "output.txt"
     ) private String output_filePath = "";
+    @CommandLine.Option(
+            names = {"-m", "--mode"},
+            description = """
+                    The action to perform:
+                    -uppercase : Transform all characters in uppercase.
+                    -lowercase : Transform all characters in lowercase.
+                    -reverse   : Reverse the input text.
+                    -find      : Return the number of times a word is found in the text (not case sensitive).
+                    -countChar : Return the number of characters in the text.
+                    -countWords: Return the number of words in the text.
+                    -replace   : Replace a word by an other one (case sensitive).
+                    -split     : Split the text into multiple files.
+                    """,
+            defaultValue = "uppercase"
+    ) private String mode;
 
     @CommandLine.Option(
             names = {"-w", "--word"},
-            description = "The word to find"
-    ) private String find_word = "";
+            description = "For the modes 'find' and 'replace': the word to find."
+    ) private String word = "";
 
     @CommandLine.Option(
             names = {"-r", "--replacement"},
-            description = "The word to replace the other one (case sensitive)"
+            description = "For the mode 'replace': the word to replace the other one (case sensitive)"
     ) private String replacement = "";
 
     @CommandLine.Option(
             names = {"-lts", "--linestosplit"},
-            description = "The number of lines before new file",
+            description = "For the mode 'split': split the text every X lines",
             defaultValue = "2"
     ) private String linesToSplit = "";
 
-    @CommandLine.Option(
-            names = {"-m", "--mode"},
-            description = "The function to apply : uppercase, lowercase, reverse, split",
-            defaultValue = "uppercase"
-    ) private String mode;
 
     public Integer call() {
         String content;
 
-        // Read file or terminal input
+        // Read file or terminal input, but not both
         if (!input_filePath.isEmpty() && input_terminalText.isEmpty()) { // Read from file
             content = FileHandler.readFile(input_filePath);
+            if(Objects.equals(output_filePath, "output.txt"))
+                output_filePath = "output_" + input_filePath; // Default name for output file
         } else if (!input_terminalText.isEmpty() && input_filePath.isEmpty()) { // Read from terminal
             content = input_terminalText;
-        } else { // input_filePath and input_terminalText are both used or none
+        } else if (!input_terminalText.isEmpty() && !input_filePath.isEmpty()) { // input file and terminal both used
             System.out.println("Choose only one input : file (-f) or text (-t) !");
             return 1;
+        } else { // none is used
+            System.out.println("Choose one input : file (-f) or text (-t) !");
+            return 1;
         }
-        if(output_filePath.isEmpty())
-            output_filePath = "output_" + input_filePath;
 
         if(content!=null) {
             switch(mode){
                 case "find":
-                    System.out.println("The word \"" + find_word + "\" appears " +
-                            TextHandler.countOccurrences(content, find_word) + " times in the text !");
-                    break;
+                    if(!word.isEmpty()) {
+                        System.out.println("The word \"" + word + "\" appears " +
+                                TextHandler.countOccurrences(content, word) + " times in the text !");
+                        break;
+                    } else {
+                        System.out.println("/!\\ This mode requires the parameters 'word' (-w) to be used !");
+                        return 1;
+                    }
                 case "countWords":
                     System.out.println("There's " + TextHandler.countWords(content) + " words in the text !");
                     break;
@@ -85,16 +102,21 @@ class Main implements Callable<Integer> {
                     }
                     break;
                 case "replace":
-                    content = TextHandler.replaceWord(content, find_word, replacement);
-                    FileHandler.writeFile(output_filePath, content);
-                    break;
+                    if (!word.isEmpty() && !replacement.isEmpty()) {
+                        content = TextHandler.replaceWord(content, word, replacement);
+                        FileHandler.writeFile(output_filePath, content);
+                        break;
+                    } else {
+                        System.out.println("/!\\ This mode requires the parameters 'word' (-w) and 'replacement' (-r)" +
+                                " to be used !");
+                        return 1;
+                    }
                 default:
-                    // Apply function to modify text
                     content = TextHandler.transform(content, mode);
-                    // Write in output file
                     FileHandler.writeFile(output_filePath, content);
                     break;
             }
+
             System.out.println("Job's done !");
             return 0;
         } else {
